requirements.txt
django==5.2.6
gunicorn
djangorestframework-simplejwt
whitenoise==6.9.0